# docs/admin/admin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 521386980 进入技术交流群 请先star，谢谢！",
        "title": "gmgo-admin API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/go-admin-team/go-admin/blob/master/LICENSE.md"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "description": "获取token\nLoginHandler can be used by clients to get a jwt token.\nPayload needs to be json in the form of {\"username\": \"USERNAME\", \"password\": \"PASSWORD\"}.\nReply will be of the form {\"token\": \"TOKEN\"}.\ndev mode：It should be noted that all fields cannot be empty, and a value of 0 can be passed in addition to the account password\n注意：开发模式：需要注意全部字段不能为空，账号密码外可以传入0值",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登陆"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"成功退出系统\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.Login": {
            "type": "object",
            "required": [
                "code",
                "password",
                "username",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}